+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2025-08-04
+
+### Added
+- ⬆️ Added `tools upgrade` command to upgrade terminal or specific tools via:
+    - `tools upgrade self`
+    - `tools upgrade <tool>@<version>`
+- 🔄 Supports automatic upgrade with version checks and exclusions
+ 
+### [Unreleased]
+- Add tools remove command to uninstall tools
 
 ## [1.3.0] - 2025-07-29